Various UI fixes & Code cleanups (#2703)
* Small editor theming fix

Character Editor had wrong BG color when no charater was selected.

Adds Icons to the Create New Character/Timeline button (when those editors are empty), to make it clearer that the buttons in the toolbar do the same thing.

* Fixes for previous commit

* Clean up event code script regions & Add some docs links

* Set Version Button Hover Cursor

* Slightly adjust home page tab styling

* Small fixes

- makes sure sidebar tree margin is correct on load
- makes key a variant, because it can be null

* Fix typo

* Make sure fields in int mode enforce 1.0 step

This was incorrect in the z-index field of the character event.

* Small fix to positioning of "2" in the logo on home tab

* Allow creating timeline/character from right click in sidebar empty area

* Slightly adjust button naming & change "import portraits from folder"

Import portraits from Folder now expects selecting a series of files instead of a folder, as it is more intuitve and flexible.

Changes various icon button tooltips (E.g. New Timeline instead of Add Timeline) to be more in line with godot terminology

* Disable auto-translate for the whole plugin interface

When having godot in a different language, random parts in dialogic where translated, resulting in a very bad look. For now auto-translation should be disabled on the whole interface.

* Cleanup/Simplify autoload adding

The autoload was re-added every time the plugin loaded, because godot now saves autoloads by UID and thus "Core" was never in the return value of the setting.
I think it's fine to drop this small auto-upgrade.

* Fix connection lines on choices

The lines on choices where broken when choices where directly inside of each other or at the end of a timeline.

* Add a setting to events to prevent them from expanding on creation

Though on most events with a body it's useful to have them open when first created, on some it's kinda annoying imo. So for now the label, choice, wait input will stay collapsed when created.
My reasoning is a bit different for them. For label and wait input I think you mostly don't need those settings. But for choices it's mostly because you will often add a bunch of choices at once and then having them all expanded is unnecessary.

* Make use of EditorInterface singleton directly

This is functionality since 4.2, so it should be fine to switch to it now.

* Slightly adjust drawing of group lines

* Improve VE line drawing, better UI/UX for empty & collapsed choices/ifs

* More attempts at cleaning up Plugin Initilization logic

* Attempt at fixing 4.4 compatibility