Adjust placement of buttons & improve label list
This places the add timeline/character and reference manager buttons in the sidebar instead of the toolbar.
Based on user testing (and the rest of godot) this is a more intuitive position for these buttons).

Also cleans up the logic for putting buttons in places.

Improves the code and UX of the label list in the sidebar:
- select event when clicking in content list
- hide content list when no labels are present
- update content list when events are moved or deleted